Made changes to custom_exceptions. 1) Moved custom exceptions from gennav.utils to gennav.utils.exceptions . 2) Changed the return statement of all the custom_exceptions to suit py2 as well as py3. 3) Renamed GoalStateinObs and StartStateinObs. 4) Made necessary changes to all file importing custom_exceptions. 5) Removed Key_Error exception from custom_exceptions. 6) Fixed lint errors. 7) (Minor): Changed buffer_dist in rrt_test file.
