+++ docs/source/gennav.utils.rst
@@ -20,14 +20,6 @@ gennav.utils.common module
    :undoc-members:
    :show-inheritance:
 
-gennav.utils.custom\_exceptions module
---------------------------------------
-
-.. automodule:: gennav.utils.custom_exceptions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 gennav.utils.geometry module
 ----------------------------
