+++ gennav/planners/prm/prm.py
@@ -1,6 +1,10 @@
 from gennav.planners.base import Planner
 from gennav.utils import Trajectory
-from gennav.utils.custom_exceptions import GoalStateinObs, PathNotFound, StartStateinObs
+from gennav.utils.exceptions.custom_exceptions import (
+    InvalidGoalState,
+    InvalidStartState,
+    PathNotFound,
+)
 from gennav.utils.geometry import compute_distance
 from gennav.utils.graph import Graph
 from gennav.utils.graph_search.astar import astar
@@ -79,11 +83,10 @@ class PRM(Planner):
         """
         # Check if start and goal states are obstacle free
         if not env.get_status(start):
-            raise StartStateinObs(start)
+            raise InvalidStartState(start, message="Start state is in obstacle.")
 
         if not env.get_status(goal):
-            raise GoalStateinObs(goal)
-
+            raise InvalidGoalState(goal, message="Goal state is in obstacle.")
         # construct graph
         graph = self.construct(env)
         # find collision free point in graph closest to start_point
@@ -108,7 +111,7 @@ class PRM(Planner):
         # perform astar search
         p = astar(graph, s, e)
         if len(p.path) == 1:
-            raise PathNotFound(p)
+            raise PathNotFound(p, message="Path contains only one state")
         else:
             traj.path.extend(p.path)
         # add end_point to path
